Implement review suggestions from PR #2
   - Change init_worker_pool to use set() instead of get_or_init() to
     error when called twice (options would otherwise be silently ignored)
   - Add #[non_exhaustive] to InitError for forward compatibility
   - Remove #[cfg(feature = "serde")] from run() methods - the feature
     doesn't actually gate any dependencies
   - Change task id type from usize to u32 for platform consistency
   - Extract send_request() to reduce monomorphisation cost
   - Update examples to use WorkerPoolOptions::new() pattern
   - Remove review comments from PR #2

+++ src/webworker/com.rs
@@ -5,8 +5,6 @@ use serde::{Deserialize, Serialize};
 /// It also indicates if errors occurred during the import.
 #[derive(Deserialize)]
 pub(super) struct PostInit {
-    // could remove the `success` boolean because it's entirely captured by the
-    // `message` field -- maybe rename it to `error`.
     /// `true` if initialization is complete, and `false` if import errors occurred.
     pub(crate) success: bool,
     /// The `message` is only set if `success` is false.
@@ -20,8 +18,7 @@ pub(super) struct PostInit {
 pub(super) struct Request {
     /// This is the internal task id, which is used to match a [`Response`]
     /// to the corresponding task.
-    // should probably be u32 or u64, doesn't depend on the bit width
-    pub(crate) id: usize,
+    pub(crate) id: u32,
     /// The name of the function to be executed by the worker.
     pub(crate) func_name: &'static str,
     /// The serialized argument to be passed to the function.
@@ -35,8 +32,7 @@ pub(super) struct Request {
 #[derive(Serialize, Deserialize)]
 pub(super) struct Response {
     /// The corresponding task id, matching the original id from the [`Request`] object.
-    // same here
-    pub(crate) id: usize,
+    pub(crate) id: u32,
     /// The response, which should only be `None` if the function could not be found.
     /// This should never be the case if the [`crate::func::WebWorkerFn`] was constructed
     /// using the [`crate::webworker!`] macro.
